feat: 당첨 횟수로 총 획득 상금 계산 기능 추가

+++ src/main/java/lotto/ui/ConsoleOutputView.java
@@ -34,7 +34,7 @@ public class ConsoleOutputView {
         System.out.println("---");
         for (Rank rank : Rank.values()) {
             int matchCount = rank.getMatchCount();
-            int prize = rank.getPrize();
+            long prize = rank.getPrize();
             String amountFormat = NumberFormat.getInstance().format(prize);
 
             int winningCount = winnings.getOrDefault(rank, 0);
